+++ src/factories/ProposalFactory.sol
@@ -124,8 +124,8 @@ contract ProposalFactory is IProposalFactory {
     {
         return _proposals[id];
     }
-    
-    function getProposalIDCounter() external view returns(uint256) {
+
+    function getProposalIDCounter() external view returns (uint) {
         return _proposalIdCounter;
     }
 }
